Send SMS now utilizes SmsManager
Created method to check for permissions and request if not granted.
Created method to handle permission request response and toggle the
"Share Results" button depending on the results.
Created method to check network connectivity and call status.
Created method to send sms message directly with SmsManager. The result
is toasted to the screen.